pyproject.toml
[tool.uv.workspace]
members = [
    "packages/stravapipe",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "requests-mock>=1.12.1",
    "ruff>=0.12.9",
    "types-pytz>=2025.2.0.20250809",
    "types-requests>=2.32.4.20250809",
]

[tool.ruff]
src = ["packages", "functions", "scripts", "packages/stravapipe"]
extend-exclude = ["*.ipynb"]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "RUF", # ruff-specific rules
    "PT",  # flake8-pytest-style
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["stravapipe"]
force-sort-within-sections = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = [
    "packages/stravapipe/tests",
]
pythonpath = [
    "packages/stravapipe/src",
]
addopts = "--import-mode=importlib"


[tool.mypy]
mypy_path = ["packages/stravapipe/src", "functions", "scripts"]
packages = ["stravapipe"]
explicit_package_bases = true
namespace_packages = true

# Ignore import-untyped errors for common untyped libraries
# disable_error_code = ["import-untyped"]

# Or if you prefer to be more specific, ignore specific modules:
[[tool.mypy.overrides]]
module = [
    "google.cloud.*",
    "google.auth.*",
    "google.api_core.*",
]
ignore_missing_imports = true
